# Refactor `rand` implementation