--- a/src/crs.jl
+++ b/src/crs.jl
@@ -126,6 +126,16 @@ The result inherits the unit of the `coords` after conversion to [`Cartesian`](@
 """
 tol(coords::CRS) = tol(convert(Cartesian, coords))
 
+# -------------
+# RAND METHODS
+# -------------
+
+Random.rand(C::Type{<:CRS}) = rand(Random.default_rng(), C)
+
+Random.rand(C::Type{<:CRS}, n::Int) = rand(Random.default_rng(), C, n)
+
+Random.rand(rng::Random.AbstractRNG, C::Type{<:CRS}, n::Int) = [rand(rng, C) for _ in 1:n]
+
 # -----------
 # IO METHODS
 # -----------