--- a/src/crs/projected.jl
+++ b/src/crs/projected.jl
@@ -39,7 +39,7 @@ Base.isapprox(coords₁::Projected{Datum}, coords₂::Projected{Datum}; kwargs..
 Base.isapprox(coords₁::Projected{Datum₁}, coords₂::Projected{Datum₂}; kwargs...) where {Datum₁,Datum₂} =
   isapprox(convert(Cartesian{Datum₁,3}, coords₁), convert(Cartesian{Datum₂,3}, coords₂); kwargs...)
 
-function Random.rand(rng::Random.AbstractRNG, ::Random.SamplerType{C}) where {C<:Projected}
+function Random.rand(rng::Random.AbstractRNG, C::Type{<:Projected})
   try
     convert(C, rand(rng, LatLon))
   catch