--- a/test/crs/rand.jl
+++ b/test/crs/rand.jl
@@ -1,76 +1,70 @@
 @testset "Random CRS" begin
   @testset "Basic" begin
-    rng = StableRNG(123)
+    randtest(Cartesian{NoDatum,1})
+    randtest(Cartesian{NoDatum,2})
+    randtest(Cartesian{NoDatum,3})
+    randtest(Cartesian2D)
+    randtest(Cartesian3D)
 
-    @test rand(rng, Cartesian{NoDatum,1}) isa Cartesian
-    @test rand(rng, Cartesian{NoDatum,2}) isa Cartesian
-    @test rand(rng, Cartesian{NoDatum,3}) isa Cartesian
-    @test rand(rng, Cartesian2D) isa Cartesian
-    @test rand(rng, Cartesian3D) isa Cartesian
+    randtest(Polar{NoDatum})
+    randtest(Polar)
 
-    @test rand(rng, Polar{NoDatum}) isa Polar
-    @test rand(rng, Polar) isa Polar
+    randtest(Cylindrical{NoDatum})
+    randtest(Cylindrical)
 
-    @test rand(rng, Cylindrical{NoDatum}) isa Cylindrical
-    @test rand(rng, Cylindrical) isa Cylindrical
-
-    @test rand(rng, Spherical{NoDatum}) isa Spherical
-    @test rand(rng, Spherical) isa Spherical
+    randtest(Spherical{NoDatum})
+    randtest(Spherical)
   end
 
   @testset "Geographic" begin
-    rng = StableRNG(123)
+    randtest(GeodeticLatLon{WGS84Latest})
+    randtest(GeodeticLatLon)
 
-    @test rand(rng, GeodeticLatLon{WGS84Latest}) isa GeodeticLatLon
-    @test rand(rng, GeodeticLatLon) isa GeodeticLatLon{WGS84Latest}
+    randtest(LatLon{WGS84Latest})
+    randtest(LatLon)
 
-    @test rand(rng, LatLon{WGS84Latest}) isa LatLon
-    @test rand(rng, LatLon) isa LatLon{WGS84Latest}
+    randtest(GeodeticLatLonAlt{WGS84Latest})
+    randtest(GeodeticLatLonAlt)
 
-    @test rand(rng, GeodeticLatLonAlt{WGS84Latest}) isa GeodeticLatLonAlt
-    @test rand(rng, GeodeticLatLonAlt) isa GeodeticLatLonAlt{WGS84Latest}
+    randtest(LatLonAlt{WGS84Latest})
+    randtest(LatLonAlt)
 
-    @test rand(rng, LatLonAlt{WGS84Latest}) isa LatLonAlt
-    @test rand(rng, LatLonAlt) isa LatLonAlt{WGS84Latest}
+    randtest(GeocentricLatLon{WGS84Latest})
+    randtest(GeocentricLatLon)
 
-    @test rand(rng, GeocentricLatLon{WGS84Latest}) isa GeocentricLatLon
-    @test rand(rng, GeocentricLatLon) isa GeocentricLatLon{WGS84Latest}
-
-    @test rand(rng, AuthalicLatLon{WGS84Latest}) isa AuthalicLatLon
-    @test rand(rng, AuthalicLatLon) isa AuthalicLatLon{WGS84Latest}
+    randtest(AuthalicLatLon{WGS84Latest})
+    randtest(AuthalicLatLon)
   end
 
   @testset "Projected" begin
-    rng = StableRNG(123)
+    randtest(Mercator{WGS84Latest})
+    randtest(Mercator)
 
-    @test rand(rng, Mercator{WGS84Latest}) isa Mercator
-    @test rand(rng, Mercator) isa Mercator{WGS84Latest}
+    randtest(WebMercator{WGS84Latest})
+    randtest(WebMercator)
 
-    @test rand(rng, WebMercator{WGS84Latest}) isa WebMercator
-    @test rand(rng, WebMercator) isa WebMercator{WGS84Latest}
+    randtest(PlateCarree{WGS84Latest})
+    randtest(PlateCarree)
 
-    @test rand(rng, PlateCarree{WGS84Latest}) isa PlateCarree
-    @test rand(rng, PlateCarree) isa PlateCarree{WGS84Latest}
+    randtest(Lambert{WGS84Latest})
+    randtest(Lambert)
 
-    @test rand(rng, Lambert{WGS84Latest}) isa Lambert
-    @test rand(rng, Lambert) isa Lambert{WGS84Latest}
+    randtest(Behrmann{WGS84Latest})
+    randtest(Behrmann)
 
-    @test rand(rng, Behrmann{WGS84Latest}) isa Behrmann
-    @test rand(rng, Behrmann) isa Behrmann{WGS84Latest}
+    randtest(GallPeters{WGS84Latest})
+    randtest(GallPeters)
 
-    @test rand(rng, GallPeters{WGS84Latest}) isa GallPeters
-    @test rand(rng, GallPeters) isa GallPeters{WGS84Latest}
+    randtest(WinkelTripel{WGS84Latest})
+    randtest(WinkelTripel)
 
-    @test rand(rng, WinkelTripel{WGS84Latest}) isa WinkelTripel
-    @test rand(rng, WinkelTripel) isa WinkelTripel{WGS84Latest}
+    randtest(Robinson{WGS84Latest})
+    randtest(Robinson)
 
-    @test rand(rng, Robinson{WGS84Latest}) isa Robinson
-    @test rand(rng, Robinson) isa Robinson{WGS84Latest}
+    randtest(OrthoNorth{WGS84Latest})
+    randtest(OrthoNorth)
 
-    @test rand(rng, OrthoNorth{WGS84Latest}) isa OrthoNorth
-    @test rand(rng, OrthoNorth) isa OrthoNorth{WGS84Latest}
-
-    @test rand(rng, OrthoSouth{WGS84Latest}) isa OrthoSouth
-    @test rand(rng, OrthoSouth) isa OrthoSouth{WGS84Latest}
+    randtest(OrthoSouth{WGS84Latest})
+    randtest(OrthoSouth)
   end
 end