This PR implements the same `rand` design that was used in Meshes.jl:
```
rand(rng, CRS)
rand(rng, CRS, n)
rand(CRS)
rand(CRS, n)
```
Reference: https://github.com/JuliaGeometry/Meshes.jl/blob/master/src/rand.jl

Only methods with `n` were implemented, as methods with `dims` are not used at the moment.

# Benchmarks
```
julia> using CoordRefSystems

julia> using BenchmarkTools

julia> using Random; Random.seed!(123);
```
## main
```
julia> @benchmark rand(LatLon)
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  20.380 ns … 52.065 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     20.973 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   21.005 ns ±  1.625 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▇▅ ▃█▂                                                      ▂
  ██▁███▁▄▄▄▄▅▇▇▆▅▆▇▆▆▆▆▆▅▅▄▅▅▅▄▅▅▅▅▃▅▃▅▃▄▅▄▄▄▅▃▄▁▄▁▅▃▃▁▃▃▄▁▄ █
  20.4 ns      Histogram: log(frequency) by time      28.8 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark rand(LatLon{WGS84Latest})
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  20.383 ns … 56.931 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     20.998 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   20.974 ns ±  1.208 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

              █                                                
  ▃▆▇▃▂▂▁▁▁▂▃▆██▃▂▁▁▁▂▁▂▁▁▁▁▁▂▁▁▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▂
  20.4 ns         Histogram: frequency by time        23.3 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark rand(LatLon, 10000)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  244.660 μs …   5.591 ms  ┊ GC (min … max): 0.00% … 94.00%
 Time  (median):     248.367 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   268.606 μs ± 160.285 μs  ┊ GC (mean ± σ):  5.24% ±  7.85%

  █▆▆▅▃▂▂▂▂▁▁▁                    ▂                             ▂
  ████████████████▇▇▇▆██▆▇▆▆▆▇▆▄▅███▆▄▅▅▄▃▁▃▄▃▅▄▄▃▃▁▁▃▄▁▃▁▄▁▄▃▄ █
  245 μs        Histogram: log(frequency) by time        417 μs <

 Memory estimate: 390.67 KiB, allocs estimate: 10002.

julia> @benchmark rand(LatLon{WGS84Latest}, 10000)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  244.646 μs …   5.193 ms  ┊ GC (min … max): 0.00% … 94.26%
 Time  (median):     248.767 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   270.059 μs ± 161.676 μs  ┊ GC (mean ± σ):  5.27% ±  7.85%

  █▆▆▅▃▂▁▁▁▁▁▁▁▁          ▁▁                                    ▂
  ███████████████▇▇▇▇▇▆▆▇▇██▇▇▆▆▆▆▄▅▆▄▅▄▅▅▃▄▅▄▄▄▅▄▁▄▅▃▁▄▄▅▄▄▅▅▆ █
  245 μs        Histogram: log(frequency) by time        448 μs <

 Memory estimate: 390.67 KiB, allocs estimate: 10002.
```
## PR
```
julia> @benchmark rand(LatLon)
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  20.300 ns … 183.101 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     21.169 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   21.627 ns ±   2.796 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▁  █▇▂      ▁                                                
  ▂█▆▄███▇▃▃▆▅▄█▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▁▂▂▂▁▂▁▂▁▁▂▁▁▁▁▁▁▂▁▁▁▂ ▃
  20.3 ns         Histogram: frequency by time           30 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark rand(LatLon{WGS84Latest})
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  20.290 ns … 76.787 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     21.207 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   21.519 ns ±  1.760 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

             ▂▄█▃▂                  ▁   ▆                      
  ▁▁▁▃▅▇▂▃▃▃▂███████▇▂▂▆▁▁▂▁▁▂▂▂▁▂▄▄█▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  20.3 ns         Histogram: frequency by time        23.8 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark rand(LatLon, 10000)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  193.690 μs … 125.468 ms  ┊ GC (min … max): 0.00% … 99.78%
 Time  (median):     194.201 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   217.306 μs ±   1.256 ms  ┊ GC (mean ± σ):  6.60% ±  3.53%

  █▃▅▁                          ▅▂▄▂                            ▁
  █████▄▅▅▅▄▅▆▄▅▅▅▅▄▃▅▁▃▅▄▇▆▆▆▄██████▇▅▆▅▄▅▄▄▄▄▃▅▅▆▆▅▄▃▅▅▅▅▅▅▅▅ █
  194 μs        Histogram: log(frequency) by time        269 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.

julia> @benchmark rand(LatLon{WGS84Latest}, 10000)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  193.568 μs … 129.592 ms  ┊ GC (min … max): 0.00% … 99.78%
 Time  (median):     194.378 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   219.642 μs ±   1.298 ms  ┊ GC (mean ± σ):  6.84% ±  3.70%

  █▅▃▃                  ▄▃▃▁                                    ▁
  ██████▆▆▆▆▇▆▆▆▄▆▆▆▇▆▇█████▆▆▅▆▆▅▆▅▆▅▆▅▅▄▅▆▅▅▅▆▃▅▄▃▅▅▃▄▄▃▄▃▄▄▅ █
  194 μs        Histogram: log(frequency) by time        295 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.
```
## Without `checklon` and `fixlat`
Disabling checks in the `LatLon` constructor makes the code faster but less safe:
```julia
Random.rand(rng::Random.AbstractRNG, ::Type{GeodeticLatLon{Datum}}) where {Datum} =
  GeodeticLatLon{Datum,Deg{Float64}}(-90° + 180° * rand(rng), -180° + 360° * rand(rng))
```  
```
julia> @benchmark rand(LatLon)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  5.881 ns … 30.045 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     5.909 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.957 ns ±  0.773 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

      █▆█▄▅                                                   
  ▃▅▇▇██████▄▃▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▂▂▂▂▂▂▂▂▂▁▁▁▁▁▂▁▁▁▁▁▁▁▂▁▂ ▃
  5.88 ns        Histogram: frequency by time        6.14 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark rand(LatLon{WGS84Latest})
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  6.026 ns … 17.728 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.047 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.066 ns ±  0.250 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

          ▂▂ ▃▁▂ ▁▁▁ ▁▁ ▂▆▇ █▅▇ ▅▄ ▆▅▆ ▇▇ ▃▁           ▂ ▁    
  ▂▂▄▁▄▇█▁██▁███▁███▁██▁███▁███▁██▁███▁██▁██▇▁▇█▆▁▇▆▁▆▇█▁█▅▂ ▅
  6.03 ns        Histogram: frequency by time        6.07 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark rand(LatLon, 10000)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  44.705 μs …  3.338 ms  ┊ GC (min … max): 0.00% … 97.88%
 Time  (median):     46.278 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   54.639 μs ± 51.722 μs  ┊ GC (mean ± σ):  4.07% ±  4.93%

  ▆█▇▆▄▃▂▂                                      ▃▂▂▃▃▃▃▂▂▁▁   ▂
  ██████████▆▇▆▅▄▆▄▆▃▅▆▆▄▆▆▄▅▄▄▄▄▄▁▁▃▁▃▃▃▄▁▃▄▄▄▅████████████▇ █
  44.7 μs      Histogram: log(frequency) by time        88 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.

julia> @benchmark rand(LatLon{WGS84Latest}, 10000)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  44.571 μs …  2.884 ms  ┊ GC (min … max): 0.00% … 97.79%
 Time  (median):     45.752 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   53.369 μs ± 45.298 μs  ┊ GC (mean ± σ):  3.63% ±  4.89%

  ▅█▆▃▂▂▁▁                                         ▃▃▃▃▂▂▁    ▂
  ████████▇▇▆▅▆▆▅▆▆▅▅▅▅▆▆▅▅▅▄▅▁▄▁▁▃▃▁▁▄▁▃▄▃▃▁▄▄▅▄▄▄████████▇█ █
  44.6 μs      Histogram: log(frequency) by time      85.2 μs <

 Memory estimate: 156.30 KiB, allocs estimate: 2.
```
### Version info
```
Julia Version 1.10.4
Commit 48d4fd48430 (2024-06-04 10:41 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 12 × Intel(R) Core(TM) i5-10400 CPU @ 2.90GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-15.0.7 (ORCJIT, skylake)
Threads: 8 default, 0 interactive, 4 GC (on 12 virtual cores)
Environment:
  JULIA_EDITOR = code
  JULIA_NUM_THREADS = 8
```